Commit some tweaks
