+++ ariadne/__init__.py
@@ -1,5 +1,5 @@
-from .executable_schema import make_executable_schema
-from .resolvers import add_resolve_functions_to_schema, default_resolver, resolve_to
+from .executable_schema import make_executable_schema, add_resolve_functions_to_schema
+from .resolvers import default_resolver, resolve_to
 from .utils import gql, start_simple_server
 from .wsgi_middleware import GraphQLMiddleware
